% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPCopula_helpers.R
\name{.geometric_mechanism}
\alias{.geometric_mechanism}
\title{Two-sided geometric mechanism}
\usage{
.geometric_mechanism(epsilon, sensitivity)
}
\arguments{
\item{epsilon}{Privacy budget.}

\item{sensitivity}{Global L1 sensitivity.}
}
\value{
Integer DP noise.
}
\description{
Draws discrete DP noise as the difference of two independent geometric random
variables. Used to perturb count data.
}
\keyword{internal}
